Build benchmarks for both rboehm and rustc_boehm
The build script uses cargo-script to handle the nastiness of compiling
and linking crates into our benchmarks. Unfortunately though, there
doesn't seem to be a way to specify a target output for cargo-script, so
for now we move from the cargo-script cache into the corresponding
(debug | release) target/ dir. It's a bit janky, but it works.

gc_bench assumes the user has a valid instance of rustc_boehm intalled
on their machine, and expects the path for this to be passed to the
build script by way of the RUSTC_BOEHM=/path/to/rustc_boehm environment
variable.